EPMRPP-37459: Code review fixes - 1
