+++ app/src/common/constants/chartModes.js
@@ -1,0 +1,14 @@
+export const CHART_MODES = {
+  LAUNCH_MODE: 'launch',
+  TIMELINE_MODE: 'timeline',
+
+  ALL_LAUNCHES: 'allLaunches',
+  LATEST_LAUNCHES: 'latestLaunches',
+
+  AREA_VIEW: 'area-spline',
+  BAR_VIEW: 'bar',
+  TABLE_VIEW: 'tableView',
+  PANEL_VIEW: 'panelView',
+  DONUT_VIEW: 'donutView',
+  PIE_VIEW: 'pie',
+};
