+++ app/src/pages/inside/dashboardItemPage/widgetWizardModal/widgetControls/investigatedPercentageOfLaunchesControls.jsx
@@ -11,9 +11,10 @@ import {
   STATS_SI_TOTAL,
   STATS_TI_TOTAL,
 } from 'common/constants/statistics';
+import { CHART_MODES } from 'common/constants/chartModes';
 import { validate } from 'common/utils';
 import { getWidgetModeOptions } from './utils/getWidgetModeOptions';
-import { ITEMS_INPUT_WIDTH, CHART_MODES } from './constants';
+import { ITEMS_INPUT_WIDTH } from './constants';
 import { FiltersControl, InputControl, TogglerControl } from './controls';
 import { WIDGET_WIZARD_FORM } from '../widgetWizardContent/wizardControlsSection/constants';
