+++ app/src/pages/inside/dashboardItemPage/widgetWizardModal/widgetControls/launchStatisticsControls.jsx
@@ -1,18 +1,18 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
+import { injectIntl, defineMessages, intlShape } from 'react-intl';
 import { connect } from 'react-redux';
 import { getFormValues, initialize } from 'redux-form';
 import { defectTypesSelector } from 'controllers/project';
 import { FieldProvider } from 'components/fields/fieldProvider';
-import { injectIntl, defineMessages, intlShape } from 'react-intl';
 import { validate } from 'common/utils';
+import { CHART_MODES } from 'common/constants/chartModes';
 import { getWidgetCriteriaOptions } from './utils/getWidgetCriteriaOptions';
 import { getWidgetModeOptions } from './utils/getWidgetModeOptions';
 import {
   LAUNCH_STATUSES_OPTIONS,
   GROUPED_DEFECT_TYPES_OPTIONS,
   ITEMS_INPUT_WIDTH,
-  CHART_MODES,
 } from './constants';
 import {
   FiltersControl,
